Adds filters for filtering courts on the main page
